bakkesmod/data/RocketStats/RocketStats_client/python/requirements.txt
obs-websocket-py==0.5.3
websocket-client==1.3.2
